+++ include/minishell.h
@@ -81,6 +81,7 @@ void    free_parser(t_parser **head);
 void    free_env_node(t_env *node);
 void    free_env(t_env *head);
 void    free_all(t_context *ctx);
+void    free_cmd(char **cmd);
 
 // main
 void	main_loop(t_context *ctx, char *line);
